Removed direct config access for isIncludingTax()
I programmed a module which relies on the Magento standard which is the tax helper being used for this check. The module allows customer tax group specific settings, thus a direct access to the config here instead of using the helper leads to a wrong display of the price inclusion for certain customers.